+++ js/settings.js
@@ -5,7 +5,7 @@
 import { store, KEYS } from './store.js';
 import { icon } from './icons.js';
 
-export const APP_VERSION = 'v24';
+export const APP_VERSION = 'v25';
 
 function todayStamp() {
   const d = new Date();